+++ src/Components/Authentication/LoginForm.vue
@@ -7,6 +7,18 @@
       <v-card-text>
         <p>
           Log into your Mindlogger account hosted at {{ $store.state.backend }}
+          <v-btn
+            icon
+            style="margin: 0px;"
+            @click="onSetBackend"
+          >
+            <v-icon
+              small
+              color="primary"
+            >
+              edit
+            </v-icon>
+          </v-btn>
         </p>
         <v-text-field
           v-model="username"
@@ -103,6 +115,7 @@ export default {
         password: this.password
       }).then((resp) => {
         this.$store.commit('setAuth', resp.data);
+        this.$router.push('/applets')
       }).catch((e) => {
         this.error = e.message === 'Request failed with status code 401' ? 'Invalid credentials' : e.message;
       });
@@ -112,7 +125,10 @@ export default {
     },
     onForgotPassword() {
       this.$emit('forgotPassword', null);
-    }
+    },
+    onSetBackend() {
+      this.$emit('setBackend', null);
+    },
   }
 
 };