+++ src/Components/Builder/Builder.vue
@@ -1,45 +1,12 @@
 <template>
-  <div>
-    <v-toolbar
-      app
-      dark
-      color="primary"
-      clipped-left
-    >
-      <v-toolbar-side-icon
-        @click="toggleDrawer"
-      />
-      <v-toolbar-title>{{ `Protocol Builder Version ${this.package.dependencies['protocol-builder']}` }}</v-toolbar-title>
-    </v-toolbar>
-
-    <v-navigation-drawer
-      v-model="drawer"
-      app
-      clipped
-    >
-      <v-list>
-        <v-list-tile
-          @click="toggleAbout"
-        >
-          <v-list-tile-action>
-            <v-icon>info</v-icon>
-          </v-list-tile-action>
-          <v-list-tile-content>
-            <v-list-tile-title>About</v-list-tile-title>
-          </v-list-tile-content>
-        </v-list-tile>
-      </v-list>
-    </v-navigation-drawer>
-
-    <v-content>
-      <About
-        v-if="aboutOpen"
-      />
-      <ProtocolBuilder
-        v-else
-      />
-    </v-content>
-  </div>
+  <v-content>
+    <About
+      v-if="aboutOpen"
+    />
+    <ProtocolBuilder
+      v-else
+    />
+  </v-content>
 </template>
 
 <script>
@@ -60,13 +27,5 @@ export default {
       package: PackageJson,
     };
   },
-  methods: {
-    toggleDrawer() {
-      this.drawer = !this.drawer;
-    },
-    toggleAbout() {
-      this.aboutOpen = !this.aboutOpen;
-    }
-  }
 }
 </script>